Search function now works, with small bug
There is a bug when the search layer gets initialized. Leaflet adds the markers again with a marker symbol it cant find. Small Hot-fix would be to swap the default marker symbol with a 1pixel large transparent png.
Other than that we can now search trough the markers name attribute. This can be changed to search for anything in the properties of a feature.